Fix folder removal only if exists

+++ copy-cap.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-rm -r ./tfc-cap-updater
+[[ -d ./tfc-cap-updater ]] && rm -r ./tfc-cap-updater
 mkdir -p ./tfc-cap-updater
 cp -r ../tfc-cap-updater/src/*.go ./tfc-cap-updater/
 cp -r ../tfc-cap-updater/src/go.* ./tfc-cap-updater/
